version 2.1: resume implemented

+++ QAd.sh
@@ -1,16 +1,17 @@
 #!/usr/bin/env bash
 #
 
+#:  <- this indicates docstring in this script.
 set -o errexit
 set -o nounset
 set -o pipefail
 
-declare -r SCRIPT_VER=2.0
+declare -r SCRIPT_VER=2.1
 declare -r AUTHOR="Alexandre Bouchard"
 declare -r GITHUB="https://github.com/abouchard-ds/QA.d/"
 
 # declare and initialize variables with deafult values
-# "" evaluate to FALSE in a -> if "$verbose"; scenario
+# "" evaluate to FALSE in a -> if [ "$verbose" ]; scenario
 username=""
 password=""
 stockfile=""
@@ -18,17 +19,18 @@ configuredClient=""
 verbose=""
 colnames=""
 keep_all=""
+resumable=""
 cleanup_missing=""
 starttime=$(date +%s)
+newline=$'\n'
+sleeptime=0.1
+unsafe=""
+tmpdir="./.downloads/"
 
 # GMT: Friday, January 1, 1971 12:00:00 AM
 period1="31536000"
 period2=$(date -d 'today 00:00:00' +%s)
 aggregator=QAd_dataset-"$(date +%Y%m%d)".csv
-newline=$'\n'
-sleeptime=0.1
-unsafe=""
-tmpdir="./.downloads/"
 
 # TODO: Trap signals(interrupts): http://man7.org/linux/man-pages/man7/signal.7.html
 #   EXIT        EXIT      0         termine correctement
@@ -39,8 +41,6 @@ tmpdir="./.downloads/"
 #   SIGTERM     TERM      15
 # trap trap_int INT               # ex.: run fonction trap_int s'il trap un CTRL+C
 
-# don't think like it's 1985, try to make it object-oriented. It simplifies
-# reading the script execution (kind of __init__ main) and code reuse.
 function usage() {
   cat <<EOF
   Quantitative Analysis Dataset Creator ${SCRIPT_VER}
@@ -57,7 +57,7 @@ function usage() {
       -k  keep individual .dat files
       -s  bypass sleep timer on downloads
       -m  cleanup missing symbols from stocklist file
-      -n  do not resume, start new
+      -r  resume if available
       -c  keep column names in individual CSV
       -h  help
       -V  verbose/debug
@@ -66,11 +66,15 @@ function usage() {
 EOF
 }
 
-# function available() {
-#  # bc was not installed on ubuntu 18.04
-# }
+
+function timestamp() {
+  date +"%Y-%m-%d %T"
+}
+
 
 function getConfiguredClient() {
+#: getConfiguredClient()
+#: search for and prioritize the tool to use
   if command -v wget &>/dev/null; then
     configuredClient="wget"
     if [ $verbose ]; then echo "configuredClient is ${configuredClient}"; fi
@@ -84,6 +88,8 @@ function getConfiguredClient() {
 }
 
 function httpGet() {
+#: httpGet() 
+#: function to download data in function of the configuredClient
   case "$configuredClient" in
     wget)  wget -qO- "$@" ;;
     curl)  curl -A curl -s "$@" ;;
@@ -91,10 +97,14 @@ function httpGet() {
 }
 
 function checkConnectivity() {
+#: checkConnectivity() 
+#: validate for internet connectivity
   httpGet github.com > /dev/null 2>&1 || { echo "Error: no active internet connection" >&2; return 1; }
 }
 
-function isEmailaddress() {
+function validemail() {
+#: validemail() stringToTest
+#: validate if a string looks like an email address
   if [[ "$1" =~ ^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,4}$ ]]; then
       return 0
   else
@@ -102,38 +112,50 @@ function isEmailaddress() {
   fi
 }
 
-# work in progress
 function generateList() {
-    i=0
-    declare -a URL_LIST
+#: generateList() -t hist -o urlfile.txt
+#: generate an associative array of urls and write the urls to a file
+#: assume the stocklist have been read to array STOCK_LIST
+  if [ $# -ne 4 ]; then
+    echo "generateList() need an url type -u and an output file -o."
+    return 1
+  else
+    while getopts ":t:o:" opt; do
+      case $opt in
+        t) local -r urltype=${OPTARG} ;;
+        o) local -r outfile=${OPTARG} ;;
+      esac
+    done
+
+    declare -A URL_LIST
     for STOCKS in "${STOCK_LIST[@]}"
     do
-        url="https://query1.finance.yahoo.com/v7/finance/download/${STOCKS}?period1=${period1}&period2=${period2}&interval=1d&events=history"
-        URL_LIST[$i]=${url}
-        i=$(($i+1))
+        local URL="https://query1.finance.yahoo.com/v7/finance/download/${STOCKS}?period1=${period1}&period2=${period2}&interval=1d&events=history"
+        URL_LIST[$STOCKS]=${URL}
+        echo ${URL} >> ${outfile}
     done
+  fi
 }
 
-# work in progress - it seems like it misses some. file is half the expected size
-# wget -q -O - --post-data ${postdata} '{}' >> ./${aggregator}
 # may be too fast for yahoo
 #function downloadParrallel() {
-#    echo ${url}_LIST | xargs -I '{}' -n 1 -P 8 { sleep 0.5; wget -q -O - --post-data ${postdata} '{}' >> ./${aggregator}; }
+# wget -q -O - --post-data ${postdata} '{}' >> ./${aggregator}
+#    echo $URL_LIST | xargs -I '{}' -n 1 -P 8 { sleep 0.5; wget -q -O - --post-data ${postdata} '{}' >> ./${aggregator}; }
 #}
 
 function download() {
-
-	index=1
+#: download()
+#: does the main job of this script. assume the stocklist have been read to array STOCK_LIST
+	local index=1
 
 	for STOCKS in "${STOCK_LIST[@]}"
 	do
-		percentage=$(bc <<< "scale=4; ($index/$arrayLen)*100")
-    clear >$(tty)
+    local filename=${STOCKS}".dat"
+		local percentage=$(bc <<< "scale=4; ($index/$arrayLen)*100")
     echo ""
 		echo -ne "Downloading ${STOCKS} data: (${percentage}%) of total completed.\r"
-		declare -l filename=${STOCKS}".dat"
-
-		url="https://query1.finance.yahoo.com/v7/finance/download/${STOCKS}?period1=${period1}&period2=${period2}&interval=1d&events=history"
+		
+		local url="https://query1.finance.yahoo.com/v7/finance/download/${STOCKS}?period1=${period1}&period2=${period2}&interval=1d&events=history"
     wget -q -O ${tmpdir}${filename} --post-data ${postdata} ${url}
 
     if [ $? -eq 0 ]; then
@@ -182,55 +204,51 @@ function download() {
 # number of args before entering getopts parsing
 if [ $# -lt 6 ]; then
     echo "Invalid number of arguments. Try qad.sh -h to get help."
+    usage
     exit 1
 fi
 
-# TODO : multiple identical options are possible (-a -a -a)
-while getopts ":u:p:l:ckmnsVh" opt; do
+# bc needed to calculate percent of progress
+command -v bc > /dev/null 2>&1
+if [ $? -eq 1 ]; then
+    echo "You need to install bc (a calculator)."
+    exit 1
+fi
+
+# TODO > check for duplicate: multiple identical options are possible (-a -a -a)
+# TODO > help -h make no sense. it cannot be triggered. placed 'usage' elsewhere for the moment
+while getopts ":u:p:l:ckmrsVh" opt; do
   case $opt in
       u)
-        if [ $verbose ]; then echo "-u was triggered, Parameter: ${OPTARG}" >&2; fi
-        if $(isEmailaddress ${OPTARG}); then
+        if $(validemail ${OPTARG}); then
           username=${OPTARG}
         else
           "Option -u requires a valid email address." >&2
           exit 1
         fi ;;
-      p)
-        if [ $verbose ]; then echo "-p was triggered, Parameter: ${OPTARG}" >&2; fi
-        password=${OPTARG} ;;
+      p) password=${OPTARG} ;;
       l)
-        if [ $verbose ]; then echo "-l was triggered, Parameter: ${OPTARG}" >&2; fi
         if [ -s ${OPTARG} ]; then
           stockfile=${OPTARG} 
         else
           "Option -l requires an existing, non-empty file." >&2
           exit 1
         fi ;;
-      c)
-        if [ $verbose ]; then echo "-c was triggered, Parameter: keep column names" >&2; fi
-        colnames=true ;;
-      k)
-        if [ $verbose ]; then echo "-k was triggered, Parameter: keep individual .dat files" >&2; fi 
-        keep_all=true ;;
-      m) 
-        if [ $verbose ]; then echo "-m was triggered, Parameter: remove missing symbols from config file" >&2; fi 
-        cleanup_missing=true ;;
-      n)
-        if [ $verbose ]; then echo "-n was triggered, Parameter: do not resume, start new" >&2; fi 
-        new=true ;;
-      s)
-        if [ $verbose ]; then echo "-s was triggered, Parameter: bypass sleep timer on downloads" >&2; fi 
-        unsafe=true ;;
-      V) 
-        if [ $verbose ]; then echo "-V was triggered, Parameter: verbose" >&2; fi 
-        verbose=true ;;
+      c) colnames=true ;;
+      k) keep_all=true ;;
+      m) cleanup_missing=true ;;
+      n) new=true ;;
+      r) resumable=true ;;
+      s) unsafe=true ;;
+      V) verbose=true ;;
       h) usage ;;
       \?)
         echo "Invalid option: -${OPTARG}. Try qad.sh -h to get help." >&2
+        usage
         exit 1 ;;
       :)
         echo "Option -${OPTARG} requires an argument." >&2
+        usage
         exit 1 ;;
   esac
 done
@@ -239,51 +257,54 @@ done
 getConfiguredClient
 checkConnectivity || exit 1
 
-# OPTION TO RESUME HERE
-logResume="./.qad_resume_${stockfile}_$(date +%Y%m%d).log"
+# RESUME FEATURE
 logError="./.qad_error.log"
+logResume="./.qad_resume_${stockfile}_$(date +%Y%m%d).log"
 
-if [ -s ${logResume} ]; then
-  resume=$(tail -n 1 ${logResume})
-  if [ $verbose ]; then echo "NOTICE: could resume from ${resume} if it was implemented."; fi
-else    
-  touch ${logResume}
-fi
-
-# PREPARE FOLDER AND FILE
-
-# Creer un dossier dans le working directory pour acceuillir les fichiers temporaires s'il n'existe pas
 mkdir -p ${tmpdir}
 
-# Regarde s'il y a deja un fichier pour aujourd'hui et le supprime
-# ici ca depend du resume feature
-if [ -f ./${aggregator} ] ; then
-    rm ./${aggregator}
-    # if [ ! $resume ]; then rm ./${aggregator}; fi
+if [ -f ${logError} ]; then 
+  rm ${logError} 
+else
+  touch ${logError}
 fi
 
-# Creer le fichier aggregation pour aujourd'hui
-touch ./${aggregator}
-if [[ ! -w ./${aggregator} ]]; then exit 1; fi
+if [ -s ${logResume} ]; then
+  if [ "$resumable" ]; then
+    resume=$(tail -n 1 ${logResume})
+    if [ $verbose ]; then 
+      echo "RESUMING : ${logResume} was found and last stock was ${resume}." 
+    fi
+    grep -Fvxf ${logResume} ${stockfile} > "${stockfile}.tmp"
+    stockfile="${stockfile}.tmp"
+  else
+    rm ${logResume} 
+    if [ -f ./${aggregator} ]; then 
+      rm ./${aggregator}
+    fi
+    touch ${logResume}
+    touch ./${aggregator}
+  fi
+else
+  touch ${logResume}
+  if [ -f ./${aggregator} ]; then 
+    rm ./${aggregator}
+  fi
+  touch ./${aggregator}
+fi
 
-touch ${logError}
-touch ./${aggregator}
-# if [[ ! -w ./${aggregator} ]]; then exit 1; fi
-
-
-# PRINT LA CONFIG
+# PRINT CONFIGURATION
 if [ $verbose ]; then
-    echo ""
-    echo "=================================================================================="
-    echo "Your dataset file is: ${aggregator}"
-    echo "Your stock file is : ${stockfile}"
-    echo "Your file contains "$(wc -l ${stockfile})" stocks"
-    echo "=================================================================================="
-    echo ""
+  echo ""
+  echo "=================================================================================="
+  echo "Your dataset file is: ${aggregator}"
+  echo "Your stock file is : ${stockfile}"
+  echo "Your file contains "$(wc -l ${stockfile})" stocks"
+  echo "=================================================================================="
+  echo ""
 fi
 
 # DOWNLOAD
-# Load le fichier stocklist dans un array
 declare -a STOCK_LIST
 readarray -t STOCK_LIST < ${stockfile}
 arrayLen=${#STOCK_LIST[@]}
@@ -291,53 +312,54 @@ arrayLen=${#STOCK_LIST[@]}
 postdata="user=${username}&password=${password}"
 # avoid script exit on download error
 set +e
+# clear >$(tty)
 download
 set -e
 
-
-# CLEANUP
+# MANAGE AGGREGATOR FILE
 if [ ! $keep_all ]; then 
-    # remove folder and files if any
-    rm -rf ${tmpdir}
+  rm -rf ${tmpdir}
 else
-    # at this point with keep_all option there is no BIGFILE
-    cat ${tmpdir}*.dat >> ./${aggregator}
+  cat ${tmpdir}*.dat >> ./${aggregator}
 fi
 
 sed -i '1s/^/SYMBOL,DATE,OPEN,HIGH,LOW,CLOSE,ADJ_CLOSE,VOLUME\n/' ./${aggregator}
 
-if [ ! -s ${logError} ]; then rm ${logError}; fi
-
+# CLEANUP
 if [ $cleanup_missing ]; then
-    mv ${stockfile} ${stockfile}.bak
-    grep -Fvxf ${logError} ${stockfile}.bak > ${stockfile}
+  mv ${stockfile} ${stockfile}.bak
+  grep -Fvxf ${logError} ${stockfile}.bak > ${stockfile}
+  if [ ! -s ${logError} ]; then 
+    rm ${logError}
+  fi
 fi
 
-# can remove since it was successfully completed
+if [ ! -s ${logError} ]; then rm ${logError}; fi
 rm ${logResume}
 
 # SUMMARY STATISTICS
 # benchmarking and optimization 
 if [ $verbose ]; then
-    filesize=$(du -sh ./${aggregator} | awk '{print $1}')
-    countline=$(cat ./${aggregator} | wc -l)
-    endtime=$(date +%s)
-    runtime=$((endtime-starttime))
-    summary1="Download completed. $newline Historical financial information for $arrayLen stocks."
-    summary2="File name is : ${aggregator} $newline File has $countline lines for a size of $filesize"
-    avgfilesize=$(ls -l ${tmpdir} | gawk '{sum += $5; n++;} END {print sum/n;}')
-    badstocks=$(wc -l ${logError})
+  filesize=$(du -sh ./${aggregator} | awk '{print $1}')
+  countline=$(cat ./${aggregator} | wc -l)
+  endtime=$(date +%s)
+  runtime=$((endtime-starttime))
+  summary1="Download completed. $newline Historical financial information for $arrayLen stocks."
+  summary2="File name is : ${aggregator} $newline File has $countline lines for a size of $filesize"
+  avgfilesize=$(ls -l ${tmpdir} | gawk '{sum += $5; n++;} END {print sum/n;}')
+  badstocks=$(wc -l ${logError})
 
-    function printsummary() {
-        echo ""
-        echo "=================================================================================="
-        echo $summary1
-        echo $summary2
-        echo "Average individual file size is : ${avgfilesize}"
-        echo "The was ${badstocks} Symbols not found on Yahoo."
-        echo "Runtime was ${runtime} seconds."
-        echo "=================================================================================="
-    }
+  function printsummary() {
+    echo ""
+    echo "=================================================================================="
+    echo $summary1
+    echo $summary2
+    echo "Average individual file size is : ${avgfilesize}"
+    echo "The was ${badstocks} Symbols not found on Yahoo."
+    echo "Runtime was ${runtime} seconds."
+    echo "=================================================================================="
+  }
 
-    printsummary
+  printsummary
+
 fi
